Rename utility inside connectors to common
